removed superfluous wrappers of components | reorganized the footer by components
