+++ src/Footer/index.js
@@ -1,7 +1,6 @@
 import React, { Component } from "react";
 import styled from "styled-components";
-import Selects from "./Selects";
-import Nav from "./Nav";
+import Service from "./Service";
 import Bottom from "./Bottom";
 
 const Footer = styled.footer`
@@ -10,21 +9,10 @@ const Footer = styled.footer`
   box-shadow: 0 -0.5px 0 rgba(72, 72, 72, 0.3);
 `;
 
-const Service = styled.div``;
-
 export default function() {
   return (
     <Footer>
-      <Service>
-        <div className="container">
-          <div className="row">
-            <div className="col-md-3">
-              <Selects />
-            </div>
-            <Nav />
-          </div>
-        </div>
-      </Service>
+      <Service />
       <Bottom />
     </Footer>
   );
