+++ src/Main/Featured/index.js
@@ -22,67 +22,65 @@ const Featured = styled.section``;
 
 export default function() {
   return (
-    <React.Fragment>
-      <Featured>
-        <div className="container">
-          <Head>
-            <Caption>Featured destinations</Caption>
-          </Head>
-          <Slider>
-            <div className="row">
-              <div className="col-md-2">
-                <Card>
-                  <Image src={paris} />
-                  <Title>
-                    <Name>Paris</Name>
-                  </Title>
-                </Card>
-              </div>
-              <div className="col-md-2">
-                <Card>
-                  <Image src={miami} />
-                  <Title>
-                    <Name>Miami</Name>
-                  </Title>
-                </Card>
-              </div>
-              <div className="col-md-2">
-                <Card>
-                  <Image src={tokyo} />
-                  <Title>
-                    <Name>Tokyo</Name>
-                  </Title>
-                </Card>
-              </div>
-              <div className="col-md-2">
-                <Card>
-                  <Image src={capeTown} />
-                  <Title>
-                    <Name>Cape town</Name>
-                  </Title>
-                </Card>
-              </div>
-              <div className="col-md-2">
-                <Card>
-                  <Image src={seoul} />
-                  <Title>
-                    <Name>Seoul</Name>
-                  </Title>
-                </Card>
-              </div>
-              <div className="col-md-2">
-                <Card>
-                  <Image src={losAngeles} />
-                  <Title>
-                    <Name>Los Angeles</Name>
-                  </Title>
-                </Card>
-              </div>
+    <Featured>
+      <div className="container">
+        <Head>
+          <Caption>Featured destinations</Caption>
+        </Head>
+        <Slider>
+          <div className="row">
+            <div className="col-md-2">
+              <Card>
+                <Image src={paris} />
+                <Title>
+                  <Name>Paris</Name>
+                </Title>
+              </Card>
             </div>
-            <SliderBtn style={{ backgroundImage: `url(${arrow})` }} />
-          </Slider>
-        </div>
-      </Featured>
-    </React.Fragment>
+            <div className="col-md-2">
+              <Card>
+                <Image src={miami} />
+                <Title>
+                  <Name>Miami</Name>
+                </Title>
+              </Card>
+            </div>
+            <div className="col-md-2">
+              <Card>
+                <Image src={tokyo} />
+                <Title>
+                  <Name>Tokyo</Name>
+                </Title>
+              </Card>
+            </div>
+            <div className="col-md-2">
+              <Card>
+                <Image src={capeTown} />
+                <Title>
+                  <Name>Cape town</Name>
+                </Title>
+              </Card>
+            </div>
+            <div className="col-md-2">
+              <Card>
+                <Image src={seoul} />
+                <Title>
+                  <Name>Seoul</Name>
+                </Title>
+              </Card>
+            </div>
+            <div className="col-md-2">
+              <Card>
+                <Image src={losAngeles} />
+                <Title>
+                  <Name>Los Angeles</Name>
+                </Title>
+              </Card>
+            </div>
+          </div>
+          <SliderBtn style={{ backgroundImage: `url(${arrow})` }} />
+        </Slider>
+      </div>
+    </Featured>
   );
 }
